W08D3 BankOfSimba: Add style.css and indexing to showall

+++ week-08/day-3/BankOfSimba/src/main/java/com/greenfoxacademy/bankofsimba/service/BankAccount.java
@@ -8,7 +8,7 @@ public class BankAccount {
   private String animalType;
   private DecimalFormat balanceFormat;
 
-  public BankAccount(String name, int balance, String animalType) {
+  public BankAccount(String name, double balance, String animalType) {
     this.name = name;
     this.balance = balance;
     this.animalType = animalType;
